@@ -1,7 +1,8 @@
 import { TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { WeatherService } from './weather.service';
 import { Forecast, Text} from '../interfaces/weather.model';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('WeatherService', () => {
     let service: WeatherService;
@@ -9,9 +10,9 @@ describe('WeatherService', () => {
 
     beforeEach(() => {
         TestBed.configureTestingModule({
-            imports: [HttpClientTestingModule],
-            providers: [WeatherService],
-        });
+    imports: [],
+    providers: [WeatherService, provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+});
         service = TestBed.inject(WeatherService);
         httpMock = TestBed.inject(HttpTestingController);
     });
